+++ apps/update.js
@@ -1,6 +1,7 @@
 import { Path, Data, file, common, UCPr } from '../components/index.js'
 import { update } from '../../other/update.js'
 import { UCPlugin } from '../model/index.js'
+import _ from 'lodash'
 
 export default class UCUpdate extends UCPlugin {
   constructor(e) {
@@ -12,6 +13,10 @@ export default class UCUpdate extends UCPlugin {
         {
           reg: /^#?UC(强制)?更新$/i,
           fnc: 'update'
+        },
+        {
+          reg: /^#?UC(强制)?更新.+/i,
+          fnc: 'updatePlugin'
         },
         {
           reg: /^#?UC全部(强制)?更新$/i,
@@ -34,11 +39,42 @@ export default class UCUpdate extends UCPlugin {
       await Update_Plugin.runUpdate(Path.Plugin_Name)
       Data.refresh()
       if (Update_Plugin.isUp) {
-        this.reply('UC-plugin更新成功，重启生效')
-        Data.execSync('pnpm i --filter=uc-plugin', Path.UC)
+        this.reply('UC-plugin更新成功，重启生效\n可使用#UC重启以前台重启云崽')
+        Data.execSync('pnpm i --filter=UC-plugin', Path.UC)
       }
     }
     return true
+  }
+
+  async updatePlugin(e) {
+    if (!this.GM) return false
+    const plugin = this.msg.match(/更新(.*)/)[1].trim()
+    const alias = file.YAMLreader(Path.get('resdata', 'pluginsAlias.yaml'))
+    let name = plugin
+    for (const info of alias) {
+      if (new RegExp(info.reg, 'i').test(plugin)) {
+        name = info.name
+        break
+      }
+    }
+    const Update_Plugin = new update()
+    Update_Plugin.e = e
+    Update_Plugin.reply = this.reply
+    if (Update_Plugin.getPlugin(name)) {
+      await Update_Plugin.runUpdate(name)
+      if (Update_Plugin.isUp) {
+        await common.sleep(2.51)
+        return this.reply('更新成功，重启生效\n可使用#UC重启 前台重启机器人')
+      }
+    } else if (name === '') {
+      await Update_Plugin.runUpdate(name)
+      if (Update_Plugin.isUp) {
+        await common.sleep(2.51)
+        return this.reply('云崽更新成功，重启生效\n可使用#UC重启 前台重启机器人')
+      }
+    } else {
+      return e.reply(`不匹配的插件名称：${name}`)
+    }
   }
 
   async updateAll(e) {
@@ -46,14 +82,14 @@ export default class UCUpdate extends UCPlugin {
     const Update_All = new update()
     Update_All.e = e
     Update_All.reply = this.reply
-    const oriVersion = UCPr.version
+    const oriDependencies = UCPr.package.dependencies
     Update_All.UCupdateAll = async function () {
       const dirs = file.readdirSync(Path.plugins, { removes: ['chatgpt-plugin', 'ji-plugin'] })
       await this.runUpdate()
       for (let plu of dirs) {
         plu = this.getPlugin(plu)
         if (plu === false) continue
-        await common.sleep(2)
+        await common.sleep(2.51)
         await this.runUpdate(plu)
       }
     }
@@ -61,11 +97,12 @@ export default class UCUpdate extends UCPlugin {
     await common.sleep(1)
     const command = this.msg.replace(/UC/i, '')
     if (Update_All.isUp) {
-      const nowVersion = UCPr.version
-      if (oriVersion !== nowVersion) {
+      const nowDependencies = UCPr.package.dependencies
+      if (!_.isEqual(oriDependencies, nowDependencies)) {
         Data.refresh()
-        Data.execSync('pnpm i --filter=uc-plugin', Path.UC)
+        Data.execSync('pnpm i --filter=UC-plugin', Path.UC)
       }
+      await common.sleep(2.51)
       return e.reply(`${command}执行成功，重启生效\n可使用#UC重启 前台重启机器人`)
     } else {
       return e.reply(`本次${command}未更新插件，无需重启`)